Remove addInvariant
