+++ src/index.ts
@@ -107,7 +107,7 @@ export * from "./api/registration/IngesterRegistration";
 export * from "./api/registration/IngesterRegistrationManager";
 export * from "./api/registration/ParametersBuilder";
 export * from "./api/registration/ParametersDefinition";
-export * from "./api/registration/ProjectInvariantRegistration";
+export * from "./pack/invariant/ProjectInvariantRegistration";
 export * from "./api/registration/PushImpactListenerRegistration";
 export * from "./api/registration/PushRegistration";
 export * from "./api/registration/ReviewListenerRegistration";
